+++ app/components/views/TicketsPage/PurchaseTab/TicketAutoBuyer/Form.js
@@ -1,215 +1,81 @@
-import { AutoBuyerSwitch, InfoModalButton, KeyBlueButton, PassphraseModalSwitch, TicketsCogs }from "buttons";
-import { TicketAutoBuyerInfoModalContent } from "modals";
-import { DcrInput, NumericInput, FeeInput } from "inputs";
-import { Tooltip } from "shared";
-import { defineMessages, FormattedMessage as T } from "react-intl";
+import { FormattedMessage as T } from "react-intl";
+import Details from "./Details";
+import QuickBar from "./QuickBar";
 import "style/StakePool.less";
-
-const messages = defineMessages({
-  balanceToMaintain: {
-    id: "autobuyer.balanceToMaintain",
-    defaultMessage: "Balance to Maintain",
-  },
-  maxFee: {
-    id: "autobuyer.maxFee",
-    defaultMessage: "Max Fee",
-  },
-  maxPriceAbsolute: {
-    id: "autobuyer.maxPriceAbsolute",
-    defaultMessage: "Max Price Absolute",
-  },
-  maxPriceRelative: {
-    id: "autobuyer.maxPriceRelative",
-    defaultMessage: "Max Price Relative",
-  },
-  maxPerBlock: {
-    id: "autobuyer.maxPerBlock",
-    defaultMessage: "Max Per Block",
-  },
-});
 
 const TicketAutoBuyerForm = ({
   isHidingDetails,
-  isTicketAutoBuyerEnabled,
-  isTicketAutoBuyerConfigDirty,
-  getTicketBuyerConfigResponse,
   balanceToMaintain,
   maxFee,
   maxPriceAbsolute,
   maxPriceRelative,
   maxPerBlock,
+
+  onStartAutoBuyer,
+  onDisableTicketAutoBuyer,
+  onToggleShowDetails,
+  isTicketAutoBuyerEnabled,
+  
+  isTicketAutoBuyerConfigDirty,
+  getTicketBuyerConfigResponse,
+  formatMessage,
   maxFeeError,
   balanceToMaintainError,
   maxPriceAbsoluteError,
   maxPriceRelativeError,
-  formatMessage,
   onChangeBalanceToMaintain,
   onChangeMaxFee,
   onChangeMaxPriceAbsolute,
   onChangeMaxPriceRelative,
   onChangeMaxPerBlock,
   onUpdateTicketAutoBuyerConfig,
-  onStartAutoBuyer,
-  onDisableTicketAutoBuyer,
-  onToggleShowDetails,
-  canNotEnableAutobuyer
+  canNotEnableAutobuyer,
 }) => (
-  <div>
-    <div className="stakepool-voting-title-area">
-      <div className="stakepool-voting-title-area-name">
-        <T id="autobuyer.title" m="Automatic Purchase" /></div>
-    </div>
-    <div className={isHidingDetails ? "stakepool-flex-height-auto-buyer-hidden" : "stakepool-flex-height-auto-buyer-shown" }>
-      <div className="stakepool-auto-buyer-row">
-      { isTicketAutoBuyerEnabled ?
-        <AutoBuyerSwitch enabled onClick={onDisableTicketAutoBuyer} /> :
-        <PassphraseModalSwitch
-          modalTitle={<T id="tickets.startAutoBuyerConfirmation" m="Start Ticket Buyer Confirmation" />}
-          onSubmit={onStartAutoBuyer}
-        />
-      }
-        <div className="stakepool-auto-buyer-quick-bar-row">
-          {isHidingDetails ? (
-            <div>
-              <Tooltip text={<T id="autobuyer.balanceToMaintain" m="Balance to Maintain" /> }>
-                <div className="stakepool-balance-to-maintain-icon">{balanceToMaintain}</div>
-              </Tooltip>
-              <Tooltip text={<T id="autobuyer.maxFee" m="Max Fee" /> }>
-                <div className="stakepool-max-fee-icon">{maxFee} DCR</div>
-              </Tooltip>
-              <Tooltip text={<T id="autobuyer.maxPriceAbsolute" m="Max Price Absolute" /> }>
-                <div className="stakepool-max-price-absolute-icon">{maxPriceAbsolute} DCR</div>
-              </Tooltip>
-              <Tooltip text={<T id="autobuyer.maxPriceRelative" m="Max Price Relative" /> }>
-                <div className="stakepool-max-price-relative-icon">{maxPriceRelative}%</div>
-              </Tooltip>
-              <Tooltip text={<T id="autobuyer.maxPerBlock" m="Max Per Block" /> }>
-                <div className="stakepool-max-per-block-icon">{maxPerBlock}</div>
-              </Tooltip>
-            </div>
-          ) : null}
-        </div>
-        <div className="stakepool-purchase-ticket-input-buttons">
-          <InfoModalButton
-            modalTitle={<h1><T id="accounts.automaticPurchaseInfo" m="Automatic Purchase Information" /></h1>}
-            modalContent={<TicketAutoBuyerInfoModalContent />}
-          />
-          <TicketsCogs opened={isHidingDetails} onClick={onToggleShowDetails} />
-        </div>
+    <Aux>
+      <div className="stakepool-voting-title-area">
+        <div className="stakepool-voting-title-area-name">
+          <T id="autobuyer.title" m="Automatic Purchase" /></div>
       </div>
       <div className={isHidingDetails ? "stakepool-flex-height-auto-buyer-hidden" : "stakepool-flex-height-auto-buyer-shown"}>
-        <div hidden={isHidingDetails ? true : false} className="stakepool-auto-buyer-advanced-area">
-          {canNotEnableAutobuyer ?
-            <div className="orange-warning">
-              <T id="autobuyer.canNotEnableAutobuyer" m="*Please fix fields with errors" />
-            </div> : null}
-          <div className="stakepool-purchase-ticket-row">
-            <div className="stakepool-purchase-ticket-row-left">
-              <div className="stakepool-auto-buyer-icon-areas-expand"><div className="stakepool-balance-to-maintain-icon">
-                {<T id="autobuyer.balanceToMaintain" m="Balance to Maintain" />}:
-            </div></div>
-              <div className="stakepool-purchase-ticket-num-input">
-                <div className="stakepool-input-form-purchase-ticket">
-                  <DcrInput
-                    placeholder={formatMessage(messages.balanceToMaintain)}
-                    value={balanceToMaintain}
-                    onChange={e => onChangeBalanceToMaintain(e.target.value)}
-                    invalid={balanceToMaintainError}
-                    invalidMessage={<T id="autobuyer.balanceToMaintainError"
-                      m="Your balance to mantain is invalid" />}
-                    showErrors
-                    required
-                  />
-                </div>
-              </div>
-            </div>
-            <div className="stakepool-purchase-ticket-row-right">
-              <div className="stakepool-auto-buyer-icon-areas-expand"><div className="stakepool-max-fee-icon">
-                {<T id="autobuyer.maxFee" m="Max Fee" />}:
-            </div></div>
-              <div className="stakepool-purchase-ticket-num-input">
-                <div className="stakepool-input-form-purchase-ticket">
-                  <FeeInput
-                    placeholder={formatMessage(messages.maxFee)}
-                    value={maxFee}
-                    onChange={e => onChangeMaxFee(e.target.value)}
-                    invalid={maxFeeError}
-                    invalidMessage={<T id="autobuyer.invalidMaxFee" m="*Invalid max fee (0 - 0.1 DCR/KB)" />}
-                    showErrors
-                    required
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="stakepool-purchase-ticket-row">
-            <div className="stakepool-purchase-ticket-row-left">
-              <div className="stakepool-auto-buyer-icon-areas-expand"><div className="stakepool-max-price-absolute-icon">
-                {<T id="autobuyer.maxPriceAbsolute" m="Max Price Absolute" />}:
-            </div></div>
-              <div className="stakepool-purchase-ticket-num-input">
-                <div className="stakepool-input-form-purchase-ticket">
-                  <DcrInput
-                    placeholder={formatMessage(messages.maxPriceAbsolute)}
-                    value={maxPriceAbsolute}
-                    onChange={e => onChangeMaxPriceAbsolute(e.target.value)}
-                    invalid={maxPriceAbsoluteError}
-                    showErrors
-                    required
-                  />
-                </div>
-              </div>
-
-            </div>
-            <div className="stakepool-purchase-ticket-row-right">
-              <div className="stakepool-auto-buyer-icon-areas-expand"><div className="stakepool-max-price-relative-icon">
-                {<T id="autobuyer.maxPriceRelative" m="Max Price Relative" />}:
-            </div></div>
-              <div className="stakepool-purchase-ticket-num-input">
-                <NumericInput
-                  placeholder={formatMessage(messages.maxPriceRelative)}
-                  value={maxPriceRelative}
-                  onChange={e => onChangeMaxPriceRelative(e.target.value)}
-                  unit="%"
-                  invalid={maxPriceRelativeError}
-                  showErrors
-                  required
-                />
-              </div>
-
-            </div>
-          </div>
-          <div className="stakepool-purchase-ticket-row">
-            <div className="stakepool-purchase-ticket-row-left">
-              <div className="stakepool-auto-buyer-icon-areas-expand"><div className="stakepool-max-per-block-icon">
-                {<T id="autobuyer.maxPerBlock" m="Max Per Block" />}:
-            </div></div>
-              <div className="stakepool-purchase-ticket-num-input">
-                <div className="stakepool-input-form-purchase-ticket">
-                  <NumericInput
-                    placeholder={formatMessage(messages.maxPerBlock)}
-                    value={maxPerBlock}
-                    onChange={e => onChangeMaxPerBlock(e.target.value)}
-                    showErrors
-                    required
-                  />
-                </div>
-              </div>
-            </div>
-          </div>
-          <div hidden={!getTicketBuyerConfigResponse}>
-            <KeyBlueButton
-              className="stakepool-content-purchase-button"
-              disabled={!isTicketAutoBuyerConfigDirty}
-              onClick={onUpdateTicketAutoBuyerConfig}
-            >
-              <T id="autobuyer.updateConfigBtn" m="Update Config" />
-            </KeyBlueButton>
-          </div>
-        </div>
+        <QuickBar {...{
+          isHidingDetails,
+          onStartAutoBuyer,
+          onDisableTicketAutoBuyer,
+          onToggleShowDetails,
+          isTicketAutoBuyerEnabled,
+          balanceToMaintain,
+          maxFee,
+          maxPriceAbsolute,
+          maxPriceRelative,
+          maxPerBlock
+        }}
+        
+        />
+        <Details {...{
+          isTicketAutoBuyerConfigDirty,
+          getTicketBuyerConfigResponse,
+          formatMessage,
+          maxFeeError,
+          balanceToMaintainError,
+          maxPriceAbsoluteError,
+          maxPriceRelativeError,
+          onChangeBalanceToMaintain,
+          onChangeMaxFee,
+          onChangeMaxPriceAbsolute,
+          onChangeMaxPriceRelative,
+          onChangeMaxPerBlock,
+          onUpdateTicketAutoBuyerConfig,
+          canNotEnableAutobuyer,
+          isHidingDetails,
+          balanceToMaintain,
+          maxFee,
+          maxPriceAbsolute,
+          maxPriceRelative,
+          maxPerBlock
+        }}
+        />
       </div>
-    </div>
-    </div>
+    </Aux>
   );
 
 export default TicketAutoBuyerForm;
